Rename driver/miopen_RedeuctionHost.hpp to driver/miopen_Reduction.hpp
